/* PR_START p--sql-update 08 */ Move `SqlSelectStatementNode` to a separate file.
