+++ metricflow/sql/optimizer/required_column_aliases.py
@@ -18,9 +18,9 @@ from metricflow.sql.sql_plan import (
     SqlPlanNodeVisitor,
     SqlSelectColumn,
     SqlSelectQueryFromClauseNode,
-    SqlSelectStatementNode,
     SqlTableNode,
 )
+from metricflow.sql.sql_select_node import SqlSelectStatementNode
 
 logger = logging.getLogger(__name__)
