+++ tests_metricflow/sql/optimizer/test_column_pruner.py
@@ -20,12 +20,11 @@ from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfi
 from metricflow.sql.optimizer.column_pruner import SqlColumnPrunerOptimizer
 from metricflow.sql.render.sql_plan_renderer import DefaultSqlPlanRenderer, SqlPlanRenderer
 from metricflow.sql.sql_plan import (
-    SqlJoinDescription,
     SqlPlanNode,
     SqlSelectColumn,
-    SqlSelectStatementNode,
     SqlTableNode,
 )
+from metricflow.sql.sql_select_node import SqlJoinDescription, SqlSelectStatementNode
 from tests_metricflow.sql.compare_sql_plan import assert_default_rendered_sql_equal
 
 logger = logging.getLogger(__name__)
