+++ tests_metricflow/sql/optimizer/test_rewriting_sub_query_reducer.py
@@ -18,12 +18,10 @@ from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfi
 
 from metricflow.sql.optimizer.rewriting_sub_query_reducer import SqlRewritingSubQueryReducer
 from metricflow.sql.sql_plan import (
-    SqlJoinDescription,
-    SqlOrderByDescription,
     SqlSelectColumn,
-    SqlSelectStatementNode,
     SqlTableNode,
 )
+from metricflow.sql.sql_select_node import SqlJoinDescription, SqlOrderByDescription, SqlSelectStatementNode
 from tests_metricflow.sql.compare_sql_plan import assert_default_rendered_sql_equal
