+++ tests_metricflow/sql/optimizer/test_table_alias_simplifier.py
@@ -15,11 +15,10 @@ from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfi
 from metricflow.sql.optimizer.table_alias_simplifier import SqlTableAliasSimplifier
 from metricflow.sql.render.sql_plan_renderer import DefaultSqlPlanRenderer, SqlPlanRenderer
 from metricflow.sql.sql_plan import (
-    SqlJoinDescription,
     SqlSelectColumn,
-    SqlSelectStatementNode,
     SqlTableNode,
 )
+from metricflow.sql.sql_select_node import SqlJoinDescription, SqlSelectStatementNode
 from tests_metricflow.sql.compare_sql_plan import assert_default_rendered_sql_equal
