+++ tests_metricflow/sql/test_render_cte.py
@@ -15,11 +15,10 @@ from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfi
 
 from metricflow.sql.sql_plan import (
     SqlCteNode,
-    SqlJoinDescription,
     SqlSelectColumn,
-    SqlSelectStatementNode,
     SqlTableNode,
 )
+from metricflow.sql.sql_select_node import SqlJoinDescription, SqlSelectStatementNode
 from tests_metricflow.sql.compare_sql_plan import assert_default_rendered_sql_equal
 
 logger = logging.getLogger(__name__)
